feat(xtask): add sequencer key rotation admin commands
