+++ xtask/Cargo.toml
@@ -55,6 +55,7 @@ serde_json.workspace = true
 serde_yaml.workspace = true
 tokio.workspace = true
 toml.workspace = true
+zeroize.workspace = true
 
 rustls = { version = "0.23", features = ["aws-lc-rs"] }
